# tpodashboard/part1/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrepVista API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/students/import/preview": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Preview Import",
        "operationId": "preview_import_students_import_preview_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_preview_import_students_import_preview_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/students/import/validate": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Validate Import",
        "operationId": "validate_import_students_import_validate_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportValidateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/students/import/commit": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "Commit Import",
        "operationId": "commit_import_students_import_commit_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportCommitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportCommitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_preview_import_students_import_preview_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_preview_import_students_import_preview_post"
      },
      "ColumnMappingSuggestion": {
        "properties": {
          "source_column": {
            "type": "string",
            "title": "Source Column"
          },
          "mapped_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mapped Field"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "sample_values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sample Values"
          }
        },
        "type": "object",
        "required": [
          "source_column",
          "mapped_field",
          "confidence",
          "sample_values"
        ],
        "title": "ColumnMappingSuggestion"
      },
...
      "DuplicateCandidate": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "register_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Register Number"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "matched_student_id": {
            "type": "string",
            "format": "uuid",
            "title": "Matched Student Id"
          },
          "match_type": {
            "type": "string",
            "title": "Match Type"
          },
          "matched_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched On"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "register_number",
          "full_name",
          "matched_student_id",
          "match_type",
          "matched_on"
        ],
        "title": "DuplicateCandidate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportCommitRequest": {
        "properties": {
          "import_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Batch Id"
          },
          "skip_duplicates": {
            "type": "boolean",
            "title": "Skip Duplicates",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "import_batch_id"
        ],
        "title": "ImportCommitRequest"
      },
      "ImportCommitResponse": {
        "properties": {
          "import_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Batch Id"
          },
          "imported_count": {
            "type": "integer",
            "title": "Imported Count"
          },
          "updated_count": {
            "type": "integer",
            "title": "Updated Count"
          },
          "ignored_count": {
            "type": "integer",
            "title": "Ignored Count"
          },
          "attention_count": {
            "type": "integer",
            "title": "Attention Count"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          }
        },
        "type": "object",
        "required": [
          "import_batch_id",
          "imported_count",
          "updated_count",
          "ignored_count",
          "attention_count",
          "total_rows"
        ],
        "title": "ImportCommitResponse"
      },
      "ImportPreviewResponse": {
        "properties": {
          "import_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Batch Id"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "detected_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Detected Columns"
          },
          "suggested_mapping": {
            "items": {
              "$ref": "#/components/schemas/ColumnMappingSuggestion"
            },
            "type": "array",
            "title": "Suggested Mapping"
          },
          "required_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Fields"
          },
          "unmapped_required_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unmapped Required Fields"
          }
        },
        "type": "object",
        "required": [
          "import_batch_id",
          "total_rows",
          "detected_columns",
          "suggested_mapping",
          "required_fields",
          "unmapped_required_fields"
        ],
        "title": "ImportPreviewResponse"
      },
      "ImportValidateRequest": {
        "properties": {
          "import_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Batch Id"
          },
          "column_mapping": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Column Mapping"
          },
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "department_id": {
            "type": "string",
            "format": "uuid",
            "title": "Department Id"
          },
          "program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Program Id"
          }
        },
        "type": "object",
        "required": [
          "import_batch_id",
          "column_mapping",
          "batch_id",
          "department_id",
          "program_id"
        ],
        "title": "ImportValidateRequest"
      },
      "ImportValidateResponse": {
        "properties": {
          "import_batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Batch Id"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "valid_rows": {
            "type": "integer",
            "title": "Valid Rows"
          },
          "invalid_rows": {
            "type": "integer",
            "title": "Invalid Rows"
          },
          "duplicate_rows": {
            "type": "integer",
            "title": "Duplicate Rows"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/RowIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "duplicates": {
            "items": {
              "$ref": "#/components/schemas/DuplicateCandidate"
            },
            "type": "array",
            "title": "Duplicates"
          },
          "ready_to_commit": {
            "type": "boolean",
            "title": "Ready To Commit"
          }
        },
        "type": "object",
        "required": [
          "import_batch_id",
          "total_rows",
          "valid_rows",
          "invalid_rows",
          "duplicate_rows",
          "issues",
          "duplicates",
          "ready_to_commit"
        ],
        "title": "ImportValidateResponse"
      },
...
      "RowIssue": {
        "properties": {
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "issue": {
            "type": "string",
            "title": "Issue"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "row_number",
          "field",
          "issue",
          "value"
        ],
        "title": "RowIssue"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}